Refactor Karata.Cards library to Karata.Pips
