+++ Karata.slnf
@@ -4,15 +4,15 @@
     "projects": [
       "src/Karata.App/Karata.App.csproj",
       "src/Karata.Bot/Karata.Bot.csproj",
-      "src/Karata.Cards/Karata.Cards.csproj",
       "src/Karata.Desktop/Karata.Desktop.csproj",
       "src/Karata.Kit/Karata.Kit.csproj",
       "src/Karata.Pebble/Karata.Pebble.csproj",
+      "src/Karata.Pips/Karata.Pips.csproj",
       "src/Karata.Server/Karata.Server.csproj",
       "src/Karata.Surface/Karata.Surface.csproj",
       "src/Karata.Web/Karata.Web.csproj",
-      "test/Karata.Cards.Tests/Karata.Cards.Tests.csproj",
-      "test/Karata.Kit.Tests/Karata.Kit.Tests.csproj"
+      "test/Karata.Kit.Tests/Karata.Kit.Tests.csproj",
+      "test/Karata.Pips.Tests/Karata.Pips.Tests.csproj"
     ]
   }
 }
