+++ Karata.slnx
@@ -11,18 +11,18 @@
     <Project Path="src/Karata.Bot/Karata.Bot.csproj" />
     <Project Path="src/Karata.BotFramework/Karata.BotFramework.csproj" />
     <Project Path="src/Karata.Browser/Karata.Browser.csproj" />
-    <Project Path="src/Karata.Cards/Karata.Cards.csproj" />
     <Project Path="src/Karata.Desktop/Karata.Desktop.csproj" />
     <Project Path="src/Karata.iOS/Karata.iOS.csproj" />
     <Project Path="src/Karata.Kit/Karata.Kit.csproj" />
     <Project Path="src/Karata.Pebble/Karata.Pebble.csproj" />
+    <Project Path="src/Karata.Pips/Karata.Pips.csproj" />
     <Project Path="src/Karata.Server/Karata.Server.csproj" />
     <Project Path="src/Karata.ServiceDefaults/Karata.ServiceDefaults.csproj" />
     <Project Path="src/Karata.Surface/Karata.Surface.csproj" />
     <Project Path="src/Karata.Web/Karata.Web.csproj" />
   </Folder>
   <Folder Name="/test/">
-    <Project Path="test/Karata.Cards.Tests/Karata.Cards.Tests.csproj" />
     <Project Path="test/Karata.Kit.Tests/Karata.Kit.Tests.csproj" />
+    <Project Path="test/Karata.Pips.Tests/Karata.Pips.Tests.csproj" />
   </Folder>
 </Solution>
