+++ src/Karata.Kit/Application/Client/State/RoomEvents.cs
@@ -1,8 +1,8 @@
 using System.Reactive;
 using System.Reactive.Linq;
 using System.Reactive.Subjects;
-using Karata.Cards;
 using Karata.Kit.Domain.Models;
+using Karata.Pips;
 
 namespace Karata.Kit.Application.Client.State;
