+++ src/Karata.Kit/Application/Store/RoomStore.cs
@@ -1,9 +1,9 @@
 using System.Collections.Immutable;
-using Karata.Cards;
 using Karata.Kit.Domain.Models;
 using Karata.Pebble;
 using Karata.Pebble.Interceptors;
 using Karata.Pebble.StateActions;
+using Karata.Pips;
 
 namespace Karata.Kit.Application.Store;
