+++ src/Karata.Kit/Application/Store/TurnStore.cs
@@ -1,8 +1,8 @@
 using System.Collections.Immutable;
-using Karata.Cards;
 using Karata.Pebble;
 using Karata.Pebble.Interceptors;
 using Karata.Pebble.StateActions;
+using Karata.Pips;
 
 namespace Karata.Kit.Application.Store;
