+++ src/Karata.Kit/Bot/Strategy/RandomValidBotStrategy.cs
@@ -1,9 +1,9 @@
 using System.Collections.Immutable;
-using Karata.Cards.Extensions;
 using Karata.Kit.Domain.Models;
 using Karata.Kit.Engine;
 using Karata.Kit.Support;
-using static Karata.Cards.Card.CardFace;
+using Karata.Pips.Extensions;
+using static Karata.Pips.Card.CardFace;
 
 namespace Karata.Kit.Bot.Strategy;
