+++ src/Karata.Kit/Domain/Models/TurnData.cs
@@ -1,4 +1,4 @@
-using Karata.Cards;
+using Karata.Pips;
 
 namespace Karata.Kit.Domain.Models;
