+++ src/Karata.Kit/Engine/Exceptions/InvalidAnswerException.cs
@@ -1,5 +1,5 @@
 using System.Collections.Immutable;
-using Karata.Cards;
+using Karata.Pips;
 
 namespace Karata.Kit.Engine.Exceptions;
