+++ src/Karata.Pips/Karata.Pips.csproj
