+++ src/Karata.Server/Models/Game.cs
@@ -1,7 +1,7 @@
 using System.ComponentModel.DataAnnotations.Schema;
 using Karata.Kit.Domain.Models;
-using static Karata.Cards.Card.CardFace;
 using static Karata.Kit.Domain.Models.CardRequestLevel;
+using static Karata.Pips.Card.CardFace;
 
 namespace Karata.Server.Models;
