+++ src/Karata.Web/Extensions/CardExtensions.cs
@@ -1,5 +1,5 @@
-using Karata.Cards;
-using static Karata.Cards.Card.CardFace;
+using Karata.Pips;
+using static Karata.Pips.Card.CardFace;
 
 namespace Karata.Web.Extensions;
