+++ src/Karata.Web/_Imports.razor
@@ -1,8 +1,8 @@
 ﻿@using System.Net.Http
 @using System.Net.Http.Json
-@using Karata.Cards
-@using Karata.Cards.Extensions
 @using Karata.Kit.Domain.Models
+@using Karata.Pips
+@using Karata.Pips.Extensions
 @using Karata.Web
 @using Karata.Web.Shared
 @using Karata.Web.Shared.Game
@@ -16,7 +16,7 @@
 @using Microsoft.JSInterop
 @using MudBlazor
 @using MudExtensions
-@using static Karata.Cards.Card
-@using static Karata.Cards.Card.CardColor
-@using static Karata.Cards.Card.CardFace
-@using static Karata.Cards.Card.CardSuit
+@using static Karata.Pips.Card
+@using static Karata.Pips.Card.CardColor
+@using static Karata.Pips.Card.CardFace
+@using static Karata.Pips.Card.CardSuit
